TasksAppWatcher: minor coroutine changes
